# Conflicts:
#	Frontend/lib/RegistrationView.dart
#	Frontend/lib/controller/RegistrationViewController.dart
#	Frontend/pubspec.yaml

+++ backend/src/main/resources/db/werwolf.sql
@@ -23,7 +23,7 @@ DROP TABLE IF EXISTS `users`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
 /*!50503 SET character_set_client = utf8mb4 */;
 CREATE TABLE `users` (
-  `id` int NOT NULL AUTO_INCREMENT,
+  `id` BIGINT NOT NULL AUTO_INCREMENT,
   `username` varchar(50) NOT NULL,
   `email` varchar(100) NOT NULL,
   `password_hash` varchar(255) NOT NULL,
